Update to 1.0.1
Updated code to version 1.0.1
Added more functions
Minor bug fixes

+++ source/KlipperSharpWebApi/KlipperSharpWebApi/Models/PrinterStatus/KlipperStatusGcodeMove.cs
@@ -7,13 +7,13 @@ namespace AndreasReitberger.Models
     {
         #region Properties
         [JsonProperty("homing_origin")]
-        public List<long> HomingOrigin { get; set; } = new();
+        public List<double> HomingOrigin { get; set; } = new();
 
         [JsonProperty("speed_factor")]
-        public long SpeedFactor { get; set; }
+        public double SpeedFactor { get; set; }
 
         [JsonProperty("gcode_position")]
-        public List<long> GcodePosition { get; set; } = new();
+        public List<double> GcodePosition { get; set; } = new();
 
         [JsonProperty("absolute_extrude")]
         public bool AbsoluteExtrude { get; set; }
@@ -22,13 +22,13 @@ namespace AndreasReitberger.Models
         public bool AbsoluteCoordinates { get; set; }
 
         [JsonProperty("position")]
-        public List<long> Position { get; set; } = new();
+        public List<double> Position { get; set; } = new();
 
         [JsonProperty("speed")]
-        public long Speed { get; set; }
+        public double Speed { get; set; }
 
         [JsonProperty("extrude_factor")]
-        public long ExtrudeFactor { get; set; }
+        public double ExtrudeFactor { get; set; }
         #endregion 
 
         #region Overrides
