+++ source/KlipperSharpWebApi/KlipperSharpWebApi/Models/PrinterStatus/KlipperStatusToolhead.cs
@@ -7,10 +7,10 @@ namespace AndreasReitberger.Models
     {
         #region Properties
         [JsonProperty("square_corner_velocity")]
-        public long SquareCornerVelocity { get; set; }
+        public double SquareCornerVelocity { get; set; }
 
         [JsonProperty("max_accel")]
-        public long MaxAccel { get; set; }
+        public double MaxAccel { get; set; }
 
         [JsonProperty("homed_axes")]
         public string HomedAxes { get; set; }
@@ -19,25 +19,25 @@ namespace AndreasReitberger.Models
         public double EstimatedPrintTime { get; set; }
 
         [JsonProperty("max_velocity")]
-        public long MaxVelocity { get; set; }
+        public double MaxVelocity { get; set; }
 
         [JsonProperty("print_time")]
         public double PrintTime { get; set; }
 
         [JsonProperty("max_accel_to_decel")]
-        public long MaxAccelToDecel { get; set; }
+        public double MaxAccelToDecel { get; set; }
 
         [JsonProperty("axis_minimum")]
-        public List<long> AxisMinimum { get; set; } = new();
+        public List<double> AxisMinimum { get; set; } = new();
 
         [JsonProperty("stalls")]
-        public long Stalls { get; set; }
+        public double Stalls { get; set; }
 
         [JsonProperty("axis_maximum")]
-        public List<long> AxisMaximum { get; set; } = new();
+        public List<double> AxisMaximum { get; set; } = new();
 
         [JsonProperty("position")]
-        public List<long> Position { get; set; } = new();
+        public List<double> Position { get; set; } = new();
 
         [JsonProperty("extruder")]
         public string Extruder { get; set; }
